Fix calculation errors.  Add seed data.

+++ db/migrate/20120103020918_create_orders.rb
@@ -8,9 +8,9 @@ class CreateOrders < ActiveRecord::Migration
       t.string :email, :null => false
       t.string :phone, :null => false
 
-      t.float :total, :null => false
-      t.float :total_due, :null => false
-      t.float :tax_charge, :null => false
+      t.decimal :total, :null => false
+      t.decimal :total_due, :null => false
+      t.decimal :tax_charge, :null => false
       t.string :status, :null => false
 
       t.timestamps
